RL Cartpole Control/requirements.txt
# pytorch. see https://pytorch.org/
#these are the packages you might need if you don't have them installed already
numpy
sklearn
pillow
torchvision
pyglet~=1.5.27  ## any version >2.0 will throw error for rendering.py